Split ProductProvenanceRetriever into to classes
ProductProvenanceLookup is the API intended for accessing the information.
ProductProvenanceRetriever now just handles the retrieving of data from storage.
