+++ DataFormats/Provenance/interface/Provenance.h
@@ -29,7 +29,7 @@ existence.
 namespace edm {
   class MergeableRunProductMetadataBase;
   class ProductProvenance;
-  class ProductProvenanceRetriever;
+  class ProductProvenanceLookup;
 
   class Provenance {
   public:
@@ -57,7 +57,7 @@ namespace edm {
     std::string const& processName() const { return stable().processName(); }
     std::string const& productInstanceName() const { return stable().productInstanceName(); }
     std::string const& friendlyClassName() const { return stable().friendlyClassName(); }
-    ProductProvenanceRetriever const* store() const { return store_; }
+    ProductProvenanceLookup const* store() const { return store_; }
     std::set<std::string> const& branchAliases() const { return stable().branchAliases(); }
 
     // Usually branchID() and originalBranchID() return exactly the same result.
@@ -74,7 +74,7 @@ namespace edm {
 
     void write(std::ostream& os) const;
 
-    void setStore(ProductProvenanceRetriever const* store) { store_ = store; }
+    void setStore(ProductProvenanceLookup const* store) { store_ = store; }
 
     ProductID const& productID() const { return stable().productID(); }
 
@@ -90,7 +90,7 @@ namespace edm {
 
   private:
     StableProvenance stableProvenance_;
-    ProductProvenanceRetriever const* store_;
+    ProductProvenanceLookup const* store_;
     MergeableRunProductMetadataBase const* mergeableRunProductMetadata_;
   };
